api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "osint-ai-agent"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
description = "OSINT AI Agent"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "langchain-core",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-google-genai",
    "tavily-python",
    "sse-starlette",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
